+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+- Add integration tests
+
 ## 1.0.0
 
 - Birth of pullie